controlplane: stop leaking in load-balancer specifics
All that these values are used for is to construct the tls_san field in
the rke2 config.

By simply asking for that list in the controlplane (and the RKE2_URL, as
we already do in the worker module), we could very well decide not to
create a load-balancer in front, or easily add a DNS hostname to the SAN
list.