controlplane_lb: init
Create a separate (sub)module to create the loadbalancer in front of the
control plane.

Contrary to how it was done in base, we ask for the subnet_id, not
network_id, as creating a loadbalancer in a network without a subnet
fails.

Also, explicitly ask for the type of load balancer, instead of
defaulting to lb11.

Rename outputs to something more self-explaining.

Drop the usage of cluster_name. Due to the label selector, we can't
deploy more than one cluster in the same project anyways.