remove base module
This only created a hcloud_network and hcloud_network_subnet. Some of
the values hidden behind variables with defaults, some hardcoded.

In fact, we might want to have different subnets for lb, controlplane
and workers, and how exactly we want it also depends on the environment
this is instantiated in. It's much simpler to do this from main.tf, as
an example.