controlplane: use terraform to render userdata
This moves most files to persist into files on disk, and removes a lot
of the error-prone multi-line yaml inserts.

We now use jq to edit the node-id field into the rke config file, and
use jq further down to determine whether we're the first controlplane or
not (by the presence of the `server` field).