clarify cloud-controller settings
`disable-cloud-controller` controls will prevent the rke2 default cloud
controller manager from installing. They both bind on port 10258 in the
host namespace, so we shouldn't deploy it when the hetzner ccm is
enabled.

For both controlplane and workers, the `hetzner_ccm_enabled` variable
now also controls if the `cloud-provider=external` kubelet arg is set
(so the CCM will start labelling nodes).