controlplane: add seed_node variable
This determines whether the first controlplane node will be a seed node.

It needs to be set to true initially (to create the cluster), and set to
false once the cluster was deployed (which will obviously redeploy the
seed node).

This is to prevent the seed node from recreating a (second) cluster on
redeploy.